moved function prototypes into instmsg.h
pull email.h out of fcns,h

add colors.h

add defaults.h

added extract.h

moved function prototypes into listplus.h

made arcs a vector

update todo

+++ bbs/valscan.cpp
@@ -25,6 +25,7 @@
 #include "bbs/datetime.h"
 #include "bbs/fcns.h"
 #include "bbs/input.h"
+#include "bbs/msgbase1.h"
 #include "bbs/read_message.h"
 #include "bbs/subxtr.h"
 #include "bbs/vars.h"
@@ -43,7 +44,7 @@ void valscan() {
     tmp_disable_conf(true);
   }
   bool done = false;
-  for (int sn = 0; sn < session()->subboards.size() && !hangup && !done; sn++) {
+  for (size_t sn = 0; sn < session()->subboards.size() && !hangup && !done; sn++) {
     if (!iscan(sn)) {
       continue;
     }
